contentfulの処理戻す
